Prometheus exporter for open census spans
We have added open census span calls in interesting parts of the code,
but without creating a jaeger or zipkin cluster they are not going
anywhere.

Here we add a Prometheus exporter that creates a Histogram for each of
the spans we have added. Every sampled invocation of a Span is
Observed by the Histogram. This commit sets opencensus to sample every
invocation of spans, instead of the default sampler of 1 in
10000(ish).

The Exporter ignores any spans that have a '/' prefix as it's be
obvserved that gin creates spans with names containing IDs. This
endpoints are metric'd via the standard method anyway.

This is a bit rough, but ready.